Seeding data for Transportation, HealthCare and LivingCosts

+++ server/db/models/Healthcare.js
@@ -6,15 +6,15 @@ const db = require("../db");
 
 const Healthcare = db.define("healthcare", {
   cost: {
-    type: Sequelize.INTEGER,
+    type: Sequelize.FLOAT,
     allowNull: false,
   },
   skill: {
-    type: Sequelize.INTEGER,
+    type: Sequelize.FLOAT,
     allowNull: false,
   },
   index: {
-    type: Sequelize.INTEGER,
+    type: Sequelize.FLOAT,
     allowNull: false,
   },
 });
